refactor(breadcrumbs): rename dropdown file to follow convention
